rename delete to disconnect
