+++ packages/@eventual/core-runtime/src/local/clients/socket-client.ts
@@ -10,7 +10,7 @@ export class LocalSocketClient implements SocketClient {
     throw new Error("Method not implemented.");
   }
 
-  public delete(_socketName: string, _connectionId: string): Promise<void> {
+  public disconnect(_socketName: string, _connectionId: string): Promise<void> {
     throw new Error("Method not implemented.");
   }
