+++ packages/@eventual/core-runtime/src/workflow/call-eventual-factory.ts
@@ -15,7 +15,7 @@ import { ExpectSignalFactory } from "./call-executors-and-factories/expect-signa
 import { SearchCallEventualFactory } from "./call-executors-and-factories/open-search-client-call.js";
 import { QueueCallEventualFactory } from "./call-executors-and-factories/queue-call.js";
 import { SendSignalEventualFactory } from "./call-executors-and-factories/send-signal-call.js";
-import { SendSocketCallEventualFactory } from "./call-executors-and-factories/send-socket-call.js";
+import { SendSocketCallEventualFactory } from "./call-executors-and-factories/socket-call.js";
 import { RegisterSignalHandlerCallFactory } from "./call-executors-and-factories/signal-handler-call.js";
 import { TaskCallEventualFactory } from "./call-executors-and-factories/task-call.js";
 import { TransactionCallEventualFactory } from "./call-executors-and-factories/transaction-call.js";
